refacto(widget): move customRangeWidget
